move dummy function to app from player
